// client/src/api/model/skillCharacteristic.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * The type of Characteristic
 */
export type SkillCharacteristic =
  (typeof SkillCharacteristic)[keyof typeof SkillCharacteristic];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SkillCharacteristic = {
  Brawn: "Brawn",
  Agility: "Agility",
  Intellect: "Intellect",
  Cunning: "Cunning",
  Willpower: "Willpower",
  Presence: "Presence",
} as const;